+++ config/app.php
@@ -15,6 +15,8 @@ return [
 
     'name' => env('APP_NAME', 'Laravel'),
 
+    'version' => env('APP_VERSION', 'v1.0.0'),
+
     /*
     |--------------------------------------------------------------------------
     | Application Environment